Parameter sensitivity analytic engine.
This analytic engine is designed to compute how machine learning metrics such as accuracy depend on variations of (hyper-) parameters such as learning rate. In this initial implementation, the engine consumes data from graph or tabular data sources. It creates a `report` object that brings together various pieces of information that can be useful to build true reports. This information includes now name of a parameter, name of a metric and list of metric values.